Python/requirements.txt
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
Flask==2.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pip==23.2.1
pycparser==2.21
python-dotenv==1.0.0
python-dotenv-vault==0.6.3
requests==2.31.0
setuptools==68.2.0
urllib3==2.0.4
Werkzeug==2.3.7
